chore(release): v18.1.2
fix: call organization API if user is authenticated

Only call Organization API if the user is authenticated.
With this fix when the user isn't authenticated, it won't raise an error when calling the API.
If something strange happens it will log it, but it won't crash the Microfrontend.

fccn/nau-technical#411