feat: use the compact version of language select
Always use the compacted version of the language selection on the
header. When Richie changes the language cookie to be 'pt' the MFEs
show any language on the header, so it is kind of strange. So now
showing the language name on the header is better, it hides the problem
and give us more space on the header, but still showning an easy way the
learner to change it.